requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
sqlalchemy==2.0.34
aiosqlite==0.20.0
qdrant-client==1.11.3
python-dotenv==1.0.1
httpx==0.27.2
pdfplumber==0.11.4